feat: renamed and fix for the issue (#32)
